requirements.txt
Flask
Flask-Cors
Flask-Bcrypt
mysql-connector-python
requests
stripe
python-dotenv

